Dumb uwcolony
This commit is just an another way of saying that im stupid.

Up to 5 units from secondary race if it can go uw and primary race has "#uwbuild" will go uwrec.

So, theres to little of units for uw and usually commander is missing because of that, also commander takes place of a land one (badcom,goodcom,bestcom - there are some comander slots).

Because of that even if entire second race will go uw and if uwbuild will be possible if secrace can be uw, commanders will be stolen.

So we need colonial system actually