Probably last commit
So I have decided to end the 'stealing' program and now start to rework everything i done to addaddtitional units,with  hopefully save of functionality.

btw currently it's in test stage with cascading bugs, but i already can make  impression about whole 'stealing' idea. + a lot of debug lines, that are in need of deletion